requirements.txt
duckduckgo_search>=6.0.0
python-nmap>=0.7.1
